Fix frontend timezone display bug
The formatLocalTime function wasn't receiving the clientTz parameter,
causing JavaScript to convert times to the browser's local timezone
instead of displaying them in the requested timezone.

This caused '9am' to display as '5am' or other incorrect times
depending on the user's browser timezone.

Fixed by:
1. Pass clientTz to formatLocalTime()
2. Use timeZone option in toLocaleString() to force correct timezone display

Now times display correctly in the timezone that was used for parsing.